fix score
