+++ Tutorials/Chap6/index.md
@@ -543,9 +543,6 @@ protected override void OnCollision(CollidableObject obj)
     // 衝突対象が自機弾だったら
     if (obj is PlayerBullet)
     {
-        // スコアを加算
-        mainNode.score += score;
-
         // 死亡時エフェクトを再生
         Parent.AddChildNode(new DeathEffect(Position));
