+++ Tutorials/Chap8/Text/Spl2.cs
@@ -1,59 +1,34 @@
 using Altseed2;
-using System.Collections.Generic;
 
 namespace Tutorial
 {
-    // メインステージのクラス
-    public class MainNode : Node
+    // 敵の基礎となるクラス
+    public class Enemy : CollidableObject
     {
 
         ...略...
 
-        // エンジンに追加された時に実行
-        protected override void OnAdded()
+        // 衝突時に実行
+        protected override void OnCollision(CollidableObject obj)
         {
-            // キャラクターノードを追加
-            AddChildNode(characterNode);
+            // 衝突対象が自機弾だったら
+            if (obj is PlayerBullet)
+            {
++               // スコアを加算
++               mainNode.score += score;
 
-            // UIを表示するノード
-            var uiNode = new Node();
+                // 死亡時エフェクトを再生
+                Parent.AddChildNode(new DeathEffect(Position));
 
-            // UIノードを追加
-            AddChildNode(uiNode);
+                // 自身を削除
+                Parent.RemoveChildNode(this);
 
-            // 背景に使用するテクスチャ
-            var backTexture = new SpriteNode();
-            // 背景のテクスチャを読み込む
-            backTexture.Texture = Texture2D.LoadStrict("Resources/Background.png");
-            // 表示位置を奥に設定
-            backTexture.ZOrder = -100;
+                // 死亡時サウンドを読み込み
+                var deathSound = Sound.LoadStrict("Resources/Explosion.wav", true);
 
-            // 背景テクスチャを追加
-            AddChildNode(backTexture);
-
-            // プレイヤーを設定
-            player = new Player(this, new Vector2F(100, 360));
-
-            // キャラクターノードにプレイヤーを追加
-            characterNode.AddChildNode(player);
-
-+           // スコアを表示するノードを設定
-+           scoreNode = new TextNode();
-+           // スコア表示に使うフォントを読み込む
-+           scoreNode.Font = Font.LoadDynamicFontStrict("Resources/GenYoMinJP-Bold.ttf", 30);
-+           // スコア表示の位置を設定
-+           scoreNode.Position = new Vector2F();
-+           // スコア表示の文字を設定
-+           scoreNode.Text = "スコア";
-+
-+           // UIノードにスコア表示ノードを追加
-+           uiNode.AddChildNode(scoreNode);
-
-            // ウェーブを初期化する
-            InitWave();
-
-            // BGMを初期化する
-            InitBGM();
+                // 死亡時サウンドを再生
+                Engine.Sound.Play(deathSound);
+            }
         }
 
         ...略...
